app/requirements.txt
asgi_lifespan
argon2_cffi
asyncio
bcrypt
black
colorama
email-validator
fastapi
httpx
isort
loguru
mypy
passlib
pymongo
pre-commit
python-dotenv
python-decouple
python-jose
python-multipart
python-slugify
pytest
pytest-asyncio
pytest-cov
pytest-xdist
trio
uvicorn
